Names Updated

+++ OnlineShopping/src/main/java/com/cg/zmart/service/ProductServiceImpl.java
@@ -14,43 +14,45 @@ public class ProductServiceImpl implements ProductService {
 
 	@Autowired
 	ProductRepository prodRepo;
+	
+
+	// FETCHING ALL PRODUCTS FROM THE DB
+	@Override
+	public List<Product> getAllProducts() {
+		return prodRepo.findAll();
+	}
+	
+	//SAVE NEW PRODUCT TO DB
+	@Override
+	public Product createProduct(Product product) {
+		return prodRepo.save(product);
+	}
+	
+	//FETCH BY ID
+	@Override
+	public Optional<Product> fetchProductById(Long id) {
+		return prodRepo.findById(id);
+	}
+	
+	//DELETE BY ID
+	@Override
+	public String destroyProduct(Long id) {
+		if(id!=null) {
+		prodRepo.deleteById(id);
+		return "Product deleted successfully";
+		}else {
+			throw new NotFoundException("Invalid Id!");
+		}
+	}
 
 	@Override
 	public Product updateProduct(Long id,Product product) {
 		return prodRepo.saveAndFlush(product);
 	}
 
-	// FETCHING ALL PRODUCTS IN THE DB
-	@Override
-	public List<Product> getAllProducts() {
-		return prodRepo.findAll();
-	}
-
-	//FETCH BY ID
-	@Override
-	public Product fetchProductById(Long id) {
-		Optional<Product> product=prodRepo.findById(id);
-		if(product.isPresent()) {
-			return product.get();
-		}else {
-			throw new NotFoundException("No product exists with the given Id");
-		}
-	}
-
-	@Override
-	public List<Product> destroyProduct(Long id) {
-		prodRepo.deleteById(id);
-		return prodRepo.findAll();
-	}
-
 	@Override
 	public List<Product> filterByCategory(String category) {
 		return prodRepo.findByProductCategory(category);
-	}
-
-	@Override
-	public Product createProduct(Product product) {
-		return prodRepo.save(product);
 	}
 
 }
